change type

+++ src/Netcorext.Auth.Core/Domain/Entities/Token.cs
@@ -8,16 +8,12 @@ public class Token : Entity
     public ResourceType ResourceType { get; set; }
     public string ResourceId { get; set; } = null!;
     public string TokenType { get; set; } = null!;
-
     public string AccessToken { get; set; } = null!;
-
     public int? ExpiresIn { get; set; }
-    public int? ExpiresAt { get; set; }
-
+    public long? ExpiresAt { get; set; }
     public string? Scope { get; set; }
-
     public string? RefreshToken { get; set; }
     public int? RefreshExpiresIn { get; set; }
-    public int? RefreshExpiresAt { get; set; }
+    public long? RefreshExpiresAt { get; set; }
     public bool Disabled { get; set; }
 }
